backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2
cryptography==42.0.0

# ML/AI Core
tensorflow==2.15.0
tensorflow-recommenders==0.7.3
torch==2.1.2
torchvision==0.16.2
pytorch-geometric==2.4.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# NLP & Transformers
transformers==4.37.0
sentence-transformers==2.3.1
tokenizers==0.15.0
huggingface-hub==0.20.2

# Recommendation Libraries
surprise==0.1
implicit==0.7.2
lightfm==1.17

# Computer Vision
opencv-python==4.9.0.80
Pillow==10.2.0
ffmpeg-python==0.2.0
scikit-image==0.22.0

# Vector Databases
pinecone-client==3.0.2
faiss-cpu==1.7.4
qdrant-client==1.7.3
chromadb==0.4.22

# Search
elasticsearch==8.11.1
opensearch-py==2.4.2

# LLM Integration
ollama==0.1.6
langchain==0.1.4
langchain-community==0.0.17

# Task Queue & Workers
celery==5.3.6
redis==5.0.1
kombu==5.3.5
flower==2.0.1

# Message Queue
aiokafka==0.10.0
confluent-kafka==2.3.0

# Caching
python-redis-lock==4.0.0
aiocache==0.12.2

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pyarrow==14.0.2
polars==0.20.3
orjson==3.9.12
msgpack==1.0.7

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.40.0
structlog==24.1.0
python-json-logger==2.0.7

# Validation & Serialization
marshmallow==3.20.2
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.1
click==8.1.7
typer==0.9.0
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.2.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Date & Time
python-dateutil==2.8.2
pytz==2024.1
arrow==1.3.0

# Image Processing
imagehash==4.3.1
colorthief==0.2.1

# API Integrations
tmdbv3api==1.9.0
beautifulsoup4==4.12.3
lxml==5.1.0
scrapy==2.11.0

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Rate Limiting
slowapi==0.1.9
limits==3.7.0

# Configuration Management
dynaconf==3.2.4
pyyaml==6.0.1
toml==0.10.2

# AWS/Cloud (optional)
boto3==1.34.28
aioboto3==12.3.0

# Kubernetes
kubernetes==29.0.0

# Performance
uvloop==0.19.0
orjson==3.9.12
ujson==5.9.0

# Background Jobs
apscheduler==3.10.4
schedule==1.2.1

# Graph Processing
networkx==3.2.1
igraph==0.11.3

# Feature Engineering
category-encoders==2.6.3
feature-engine==1.6.2

# Model Serving
tensorflow-serving-api==2.15.0
tritonclient[all]==2.42.0

# Experiment Tracking
mlflow==2.9.2
wandb==0.16.2

# Data Validation
great-expectations==0.18.8
pandera==0.18.0
